refactor(territory): change AdministrativeLevel id type to Integer; update repository accordingly
feat(territory): add AdministrativeLevelService with create and delete operations

+++ territory/src/main/java/pt/estga/territory/repositories/AdministrativeLevelRepository.java
@@ -6,7 +6,7 @@ import pt.estga.territory.entities.Country;
 
 import java.util.Optional;
 
-public interface AdministrativeLevelRepository extends BaseRepository<AdministrativeLevel, Long> {
+public interface AdministrativeLevelRepository extends BaseRepository<AdministrativeLevel, Integer> {
 
 	Optional<AdministrativeLevel> findByCountryAndOsmLevel(Country country, int osmLevel);
